+++ @xen-orchestra/lite/src/components/host/system/HostSystemResourceManagement.vue
@@ -13,7 +13,7 @@ import { useControlDomainStore } from '@/stores/xen-api/control-domain.store.ts'
 import VtsQuickInfoRow from '@core/components/quick-info-row/VtsQuickInfoRow.vue'
 import UiCard from '@core/components/ui/card/UiCard.vue'
 import UiTitle from '@core/components/ui/title/UiTitle.vue'
-import { formatSizeRaw } from '@core/utils/size.util.ts'
+import { formatSize } from '@core/utils/size.util.ts'
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
 
@@ -32,8 +32,6 @@ const controllerMemory = computed(() => {
     return
   }
 
-  const controlDomainMemory = formatSizeRaw(controlDomain.memory_dynamic_max, 2)
-
-  return `${controlDomainMemory.value} ${controlDomainMemory.prefix}`
+  return formatSize(controlDomain.memory_dynamic_max, 2)
 })
 </script>